Add functions s.t. when we lats and lons, you will get out a
list of the vertices and edges in the transect ccreated from the
algorithm. Then you have to convert this information into a netcdf file
and geojson file (so you have a complete mask, not just the vertices and
edges on the border of the mask). THEN, you must process this data again to remove any duplicate vertices... There is probably a way (based on how the condition statement is written in xr_sorted_transect_edges_and_vertices) to avoid this third processing step and consider this when you first identify the edges and vertices in the first step.