Add calculations for transport through the transect surrounding a region
for all density classes (not just a single density class). I'm not
convinced this is 100% correct, or the most efficient way to do things
because it involves stacking coordinates. Next step is to quantify the
total sum of the transport, I suppose? To verify that we would expect to
see this amount of export and import at these density classes (1024-1028
kg/m^3; water lighter than that is excluded but present in the
datset)...