Cleaned up environment file and moved to root
A straight-up `conda env create --file ci/lanl_env.yml` command failed
because of some inconsistency dependencies, so I decided to create
the environment from scratch with a minimal set of conda installs.

I've imposed xgcm >= 0.9.0 to minimize growing pains later on,
when we merge some of this code with sectionate/regionate/xwmb.

Also, this now makes an editable install of the `mpasregions` package
that lives at the root level of this repo.

Added docstrings to some minor utility functions