Update sorted_transect_edges_and_vertices function
to include a break line so that islands not connected to the transect
are excluded. Add the plot created from calling this function to github
(it takes a while to plot). The .nc and .geojson files created have the
prefix SO_region_test1_transect_from_alg.` Next steps: convert open_transect_from_alg
function in the .ipynb to a function in sections.py.