Add test nb for sorting. Update sections
sections.py now accounts for edgesOnVertex, which can have np.nans
(since 2 or 3 edges on vertex is possible). These nans are stored as 0s
in edgesOnVertex (stored as n_IDs). When we convert from n_IDs to
xr_indices, we would have this value be -1. In calculate_transects, this
would make the last option in xr_edgesOnStartVertex array available to
be chosen. Fixed this by adding an intermediate step that removes
n_edgesOnStartVertex that are less than or equal to 0 (i.e., represent
nans).

Note: I only fixed this in the calculate_transects function. We may need
to consider doing this for all functions containing edgesOnVertex. We do
not need to do this for verticesOnEdge because there are no vertices
that need to be represented by nans (every edge has exactly 2 vertices
it connects)

sections.py now also removes duplicate points (target_start_pt = target_end_pt and is double-counted in individual segment builds as the last pt in the previous segment and first pt in the next) for calculate_transects_multiple_pts.