Undo prevent adding used pts to new segment
calculate_transects_multiple_pts originally allowed pts in previous
segments to be reused in new segments.

then I changed it prevent this from happening (used_vertices didn't
start empty for a new segment; it included used_vertices from previous
transects). but this caused problems with trying to reach the start/end
point.

so I've changed it back to the original to allow pts in previous
segments to be reused in new segments. Reusing pts (ie the total
transect crosses over itself) is fine b/c the
find_and_sort_transect_edges_and_vertics function ~is supposed to (check
this in tests)~ remove repeated pts, so it doesn't matter if the
transect crosses over itself. that point will likely be removed anyways
in this function...?