Add nb with transect crossing over itself
I am trying to force a transect to cross over itself by traveling from
vertex 8-6-4-6-3-5-7-9. The repeat of 6-4-6 occurs because I force the
transect segments to travel from 6-4 and back.

My problem is that when I use the
find_and_sort_transect_edges_and_vertices function, it requires choosing
a single xr_nextEdge. But since vertex 6 is connected to edge 6 (6-4
segment) and to edge 5 (6-3 segment), there are two potential
xr_nextEdges, which breaks the code.

I thought a solution would be to remove all the vertices and edges
between the first time vertex 6 is crossed and the next time it is
crossed (so vertices 4-6 and edge 6 are removed from the remaining_edges
list), but I can't seem to get this modification of the code correct.
Something is wrong with my array sizes or indexing, I think?

So 1) I don't know if my proposed solution is the right way to go about
fixing this problem, 2) My pursuit of this proposed solution is
incorrect, and 3) this solution prevents a transect from ever crossing
over itself, and should we expect that out of a region?

Note to self: You cannot fix this while making the transect algorithm
(ie prevent going to previous vertices because the calculate_transect
algorithm will try to get you as close as possible to the
target_end_point, maximizing the region area).

Also, this problem should only arise when using a calculate_transect
algorithm because why would a human choose a region where the area is
clearly zero?