Clean nb to include test definitions
Add consideration for cases where a) the transect crosses over itself
and b) when the transect follows an entire line of latitude s.t. it
crosses over a periodic boundary

Next steps:
check how to create a transect that follows a line of constant latitude,
crossing over a periodic boundary... Is it even possible to define a
mask like this?

Identify at which point in this process it doesn't matter if you created
a region mask from a .geojson file or a transect algorithm. Check if
functions used prior to that point work for from a .geojson file.