Update sorted_transect_edges_and_vertices function
Now includes consideration for if a mask uses multiple transect segments
that are not attached to each other (e.g., if cells are bounded above
and below by lines of constant latitude around the entire sphere).

Updated function will look for the vertex closest to the current vertex
and move to that (instead of looking for what vertex is attached to the
edgeOnVertex(currentVertex)).

Some things to consider:
1) is this the best way to switch transects? One thing it does not
   guarantee is that the upper and lower transects may not move in the
same direction... Should we look for the next point closest to the
starting vertex instead of the currentVertex?
2) I don't know if there's a way to use the transect algorithm + create
   mask algorithm to get transects like this... Actually, I think maybe
I did do it? The edges used to create the transect that are actually
inside the mask will get removed with the find algorithm because they
border 2 cells that are inside the mask, not 1.